refactor(elysia): nest route modules, one file per endpoint
Restructures src/routes-elysia/{health,oraclenet,dashboard}.ts into
per-module directories with one file per endpoint, mirroring the
tests/http/ layout:

  <module>/
    index.ts          # composes sub-app with prefix
    model.ts          # TypeBox schemas (shared within module)
    <endpoint>.ts     # one per endpoint

Elysia infers handler context from the `query:` schema option — explicit
`{ query: Static<...>; set: ... }` annotations fight inference, so drop
them and let Elysia do its job. Schemas still imported from model.ts via
the `query:` route option (runtime validation + inferred TS types).

All 12 routes preserved (health=3, oraclenet=4, dashboard=5). Each file
5–56 lines; well under the 250-line cap.
