+++ src/routes-elysia/dashboard/summary.ts
@@ -1,0 +1,6 @@
+import { Elysia } from 'elysia';
+import { handleDashboardSummary } from '../../server/dashboard.ts';
+
+export const summaryEndpoint = new Elysia()
+  .get('/dashboard', () => handleDashboardSummary())
+  .get('/dashboard/summary', () => handleDashboardSummary());
