+++ src/routes-elysia/health/model.ts
@@ -1,0 +1,5 @@
+import { t } from 'elysia';
+
+export const OraclesQuery = t.Object({
+  hours: t.Optional(t.String()),
+});
